セカンドコミット
編集機能を追加しました。

+++ app/controllers/tweets_controller.rb
@@ -17,12 +17,29 @@ class TweetsController < ApplicationController
   def destroy
     tweet = Tweet.find(params[:id])
     if tweet.user_id == current_user.id
-      tweet.destroy
+       tweet.destroy if tweet.user_id == current_user.id
     end
   end
 
+  def edit
+    @tweet = Tweet.find(params[:id])
+  end
 
-  private
+  def update
+    tweet = Tweet.find(params[:id])
+    if tweet.user_id == current_user.id
+      tweet.update(tweet_params)
+    end
+  end
+
+  def show
+    @tweet = Tweet.find(params[:id])
+  end
+  
+end
+
+
+private
   def tweet_params
     params.permit(:image, :text)
   end
@@ -30,5 +47,3 @@ class TweetsController < ApplicationController
   def move_to_index
     redirect_to action: :index unless user_signed_in?
   end
-
-end
